static-site: Split config used by gatsby-config into its own file
A future commit will add JSX elements to configuration, which is
something that cannot be imported by gatsby-config (it does not allow
imports to include JSX).

The best solution I can think of is to split the configuration into two
files, one (GatsbyConfig) which can be imported by gatsby-config and the
other (SiteConfig) which can include JSX elements.

Convert SiteConfig to ESM now that it is possible.

Note that GatsbyConfig isn't strictly used by gatsby-config, for example
pathPrefix is also used in a JSX page.
