+++ static-site/data/GatsbyConfig.js
@@ -1,0 +1,4 @@
+module.exports = {
+    pathPrefix: "/",
+    progressBarColor: "#5097BA"
+};
